backend/payment_schedule_api/requirements.txt
fastapi
uvicorn[standard]
pydantic
asyncpg  # Driver asíncrono para PostgreSQL
python-dotenv # Para cargar variables de .env